% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_validation.R
\name{assert}
\alias{assert}
\title{The same as stopifnot but with message.}
\usage{
assert(cond, fail_msg = NULL, ...)
}
\arguments{
\item{cond}{Condition to be evaluated}

\item{msg}{Message for stop if condition is FALSE. If not passed \code{cond} code itself is used for the message.}
}
\description{
The same as stopifnot but with message.
}
