% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_validation.R
\name{dt_validate}
\alias{dt_validate}
\title{Performes validation of data.table passed}
\usage{
dt_validate(dt, dt_arg, expected_col_class)
}
\arguments{
\item{dt}{data.table to be validated}

\item{dt_arg}{name of data.table argument for proper message content}

\item{col_class}{named list of columns and there expected classes}
}
\description{
Performes validation of data.table passed
}
\examples{
{
 dt <- data.table(c1= c(1,2,3), c2=c(Sys.Date()), c3=c("Ala & kotek"))
 dt_validate(dt, "dt", list(c2 = "Date", c1 = "numeric"))
 }

}
