% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_validation.R
\name{dt_validate_and_groom}
\alias{dt_validate_and_groom}
\title{Performes validation and grooming of data.table passed}
\usage{
dt_validate_and_groom(dt, dt_arg, expected_col_class)
}
\arguments{
\item{dt}{data.table to be validated and groomed}

\item{dt_arg}{name of data.table argument for proper message content}

\item{col_class}{named list of columns and there expected classes}
}
\value{
data.table with only expected columns in order specified by \code{col_class}.
}
\description{
Performes validation and grooming of data.table passed
}
\examples{
{
 dt <- data.table(c1= c(1,2,3), c2=c(Sys.Date()), c3=c("Ala & kotek"))
 dt_validate_and_groom(dt, "dt", list(c2 = "Date", c1 = "numeric"))
 # should output
 #            c2 c1
 # 1: 2016-04-05  1
 # 2: 2016-04-05  2
 # 3: 2016-04-05  3
}

}
