% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_forecast.R
\name{forecast_paths}
\alias{forecast_paths}
\title{Evaluate future interval-root paths}
\usage{
forecast_paths(object, ...)
}
\arguments{
\item{object}{A fitted RQR object.}

\item{...}{Method arguments.}
}
\value{
A model-specific future interval-root evaluation.
}
\description{
Evaluate future interval-root paths
}
