% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_dlm_fit.R
\name{print.rqr_dlm_mcmc}
\alias{print.rqr_dlm_mcmc}
\title{Print an RQR-DLM MCMC summary}
\usage{
\method{print}{rqr_dlm_mcmc}(x, ...)
}
\arguments{
\item{x}{An \code{rqr_dlm_mcmc} fit.}

\item{...}{Reserved; supplying an argument is an error.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an RQR-DLM MCMC summary
}
