% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_utils.R
\name{rqr_check_loss}
\alias{rqr_check_loss}
\title{RQR check loss}
\usage{
rqr_check_loss(u, coverage_level)
}
\arguments{
\item{u}{Numeric residual product.}

\item{coverage_level}{Interval coverage level.}
}
\value{
Numeric vector of RQR check losses.
}
\description{
RQR check loss
}
