% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_utils.R
\name{rqr_constants}
\alias{rqr_constants}
\title{RQR loss constants}
\usage{
rqr_constants(coverage_level, learning_rate = 1)
}
\arguments{
\item{coverage_level}{Interval coverage level in \verb{(0, 1)}.}

\item{learning_rate}{Positive generalized-Bayes learning rate.}
}
\value{
A named list with \code{alpha}, \code{omega}, \code{sigma}, \code{xi}, and \code{phi}.
}
\description{
RQR loss constants
}
