% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_desn_fit.R
\name{rqr_desn_continue}
\alias{rqr_desn_continue}
\title{Continue a frozen-design ordinary RQR-DESN chain}
\usage{
rqr_desn_continue(object, ...)
}
\arguments{
\item{object}{An \code{rqr_desn_fit} with an MCMC readout.}

\item{...}{Arguments passed to \code{\link[=rqr_mcmc_continue]{rqr_mcmc_continue()}}.}
}
\value{
A new \code{rqr_desn_fit} segment.
}
\description{
Continue a frozen-design ordinary RQR-DESN chain
}
