% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rqr_ffbs_cpp}
\alias{rqr_ffbs_cpp}
\title{C++ scalar-observation FFBS kernel}
\usage{
rqr_ffbs_cpp(
  z,
  H,
  V,
  GG,
  m0,
  C0,
  evolution_mode,
  W,
  D,
  sample_path,
  jitter_ladder,
  canonical_shift,
  evolution_label,
  allow_covariance_repair
)
}
\description{
C++ scalar-observation FFBS kernel
}
\keyword{internal}
