% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rqr_mvn_draw_cpp}
\alias{rqr_mvn_draw_cpp}
\title{C++ Gaussian covariance-draw diagnostic kernel}
\usage{
rqr_mvn_draw_cpp(mean, covariance, jitter_ladder, allow_repair)
}
\description{
C++ Gaussian covariance-draw diagnostic kernel
}
\keyword{internal}
