% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_dlm_model.R
\name{rqr_polytrend}
\alias{rqr_polytrend}
\title{Polynomial-trend RQR-DLM component}
\usage{
rqr_polytrend(order = 1L, m0 = NULL, C0 = NULL, name = "trend")
}
\arguments{
\item{order}{Positive polynomial order.}

\item{m0}{Prior state mean.}

\item{C0}{Prior state covariance.}

\item{name}{Component name.}
}
\value{
An \code{rqr_dlm_model}.
}
\description{
Polynomial-trend RQR-DLM component
}
