% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_utils.R
\name{rqr_pseudo_residual}
\alias{rqr_pseudo_residual}
\title{RQR pseudo residual from the transformed AL representation}
\usage{
rqr_pseudo_residual(y, eta1, eta2)
}
\arguments{
\item{y}{Response vector.}

\item{eta1, eta2}{Numeric root ordinates.}
}
\value{
Numeric vector \code{y^2 - y * (eta1 + eta2) + eta1 * eta2}.
}
\description{
RQR pseudo residual from the transformed AL representation
}
