% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_utils.R
\name{predict_interval}
\alias{predict_interval}
\title{Predict RQR intervals}
\usage{
predict_interval(object, ...)
}
\arguments{
\item{object}{An RQR fit object.}

\item{...}{Method-specific arguments.}
}
\value{
A list of interval draws and summaries.
}
\description{
Predict RQR intervals
}
