% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_utils.R
\name{rqr_order_endpoints}
\alias{rqr_order_endpoints}
\title{RQR ordered endpoints}
\usage{
rqr_order_endpoints(eta1, eta2)
}
\arguments{
\item{eta1, eta2}{Numeric root ordinates.}
}
\value{
A list with `lower`, `upper`, `midpoint`, and `width`.
}
\description{
RQR ordered endpoints
}
