% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_dlm_model.R
\name{rqr_regression}
\alias{rqr_regression}
\title{Regression-state RQR-DLM component}
\usage{
rqr_regression(X, m0 = NULL, C0 = NULL, name = "regression")
}
\arguments{
\item{X}{Time-by-regressor design matrix.}

\item{m0}{Prior state mean.}

\item{C0}{Prior state covariance.}

\item{name}{Component name.}
}
\value{
An `rqr_dlm_model`.
}
\description{
Regression-state RQR-DLM component
}
