% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_utils.R
\name{rqr_residual_product}
\alias{rqr_residual_product}
\title{RQR root residual product}
\usage{
rqr_residual_product(y, eta1, eta2)
}
\arguments{
\item{y}{Response vector.}

\item{eta1, eta2}{Numeric root ordinates.}
}
\value{
Numeric vector `(y - eta1) * (y - eta2)`.
}
\description{
RQR root residual product
}
