% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqr_dlm_model.R
\name{rqr_seasonal}
\alias{rqr_seasonal}
\title{Fourier seasonal RQR-DLM component}
\usage{
rqr_seasonal(period, harmonics = 1L, m0 = NULL, C0 = NULL, name = "seasonal")
}
\arguments{
\item{period}{Positive seasonal period.}

\item{harmonics}{Integer harmonics.}

\item{m0}{Prior state mean.}

\item{C0}{Prior state covariance.}

\item{name}{Component name.}
}
\value{
An \code{rqr_dlm_model}.
}
\description{
Fourier seasonal RQR-DLM component
}
