% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqrgibbs-package.R
\docType{package}
\name{rqrgibbs-package}
\alias{rqrgibbs}
\alias{rqrgibbs-package}
\title{rqrgibbs: generalized Bayes interval regression}
\description{
The package implements fixed-design and dynamic samplers for the RQR
residual-product loss. Its pseudo-AL representation is a computational
augmentation of a loss kernel, not a response likelihood.
}
\author{
\strong{Maintainer}: Antonio De Leon \email{jaguir26@ucsc.edu}

Authors:
\itemize{
  \item Antonio De Leon \email{jaguir26@ucsc.edu}
  \item Raquel Prado
  \item Bruno Sanso
}

}
\keyword{internal}
